requirements.txt
apipkg==1.5
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-cors-headers==3.2.1
django-dramatiq==0.9.1
django-environ==0.4.5
django-filter==2.2.0
django-grappelli==2.14.1
django-json-widget==1.0.0
django-object-actions==2.0.0
django-redis==4.12.1
django-reversion==3.0.7
django-reversion-compare==0.12.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dramatiq==1.8.1
drf-yasg==1.17.1
entrypoints==0.3
et-xmlfile==1.0.1
execnet==1.8.0
filemagic==1.6
django-reversion==3.0.7
django-reversion-compare==0.12.2
django-redis==4.12.1
XlsxWriter==1.3.6
sentry-sdk==0.16.3
gitpython==3.1.7
gevent==20.6.2
pika
lxml==4.6.3

# Testing

flake8==3.7.9
future==0.18.2
gevent==20.6.2
gitdb==4.0.5
GitPython==3.1.7
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
model-bakery==1.1.0
more-itertools==8.6.0
numpy==1.20.0
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.9
pandas==1.2.1
parameterized==0.7.1
pika==1.1.0
Pillow==7.1.2
pluggy==0.13.1
prometheus-client==0.9.0
psycopg2-binary==2.8.4
py==1.10.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-flake8==1.0.4
pytest-forked==1.3.0
pytest-xdist==1.31.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.4.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sentry-sdk==0.16.3
six==1.15.0
smmap==3.0.5
sqlparse==0.4.1
tablib==1.1.0
tabulate==0.8.7
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.6
xlwt==1.3.0
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.2.0
